# Optimize ORM usage, db_session instantiation, and tuning