--- a/src/dispatch/decorators.py
+++ b/src/dispatch/decorators.py
@@ -1,8 +1,12 @@
+from contextlib import _GeneratorContextManager, contextmanager
 from functools import wraps
-from typing import Any, List
+from typing import Any, Callable, List
 import inspect
 import logging
 import time
+
+from sqlalchemy.engine import Engine
+from sqlalchemy.orm import scoped_session, Session
 
 from dispatch.metrics import provider as metrics_provider
 from dispatch.organization import service as organization_service
@@ -19,7 +23,75 @@ def fullname(o):
     return f"{module.__name__}.{o.__qualname__}"
 
 
-def scheduled_project_task(func):
+@contextmanager
+def _session(
+    engine: Engine,
+    is_scoped: bool = False,
+) -> None:
+    """Provide a transactional scope around a series of operations."""
+    session = (
+        sessionmaker(bind=engine)()
+        if not is_scoped
+        else scoped_session(sessionmaker(bind=engine))()
+    )
+    try:
+        yield session
+        session.commit()
+    except Exception:
+        session.rollback()
+        raise Exception from None
+    finally:
+        session.close()
+
+
+def _execute_task_in_project_context(
+    func: Callable,
+    is_scoped: bool = False,
+    *args,
+    **kwargs,
+) -> None:
+    db_session = SessionLocal()
+    metrics_provider.counter("function.call.counter", tags={"function": fullname(func)})
+    start = time.perf_counter()
+
+    def __execute_task_within_session_context(
+        schema_session: _GeneratorContextManager[Session],
+    ) -> None:
+        kwargs["db_session"] = schema_session
+        for project in project_service.get_all(db_session=schema_session):
+            project = schema_session.merge(project)
+            kwargs["project"] = project
+            try:
+                func(*args, **kwargs)
+            except Exception as e:
+                log.exception(e)
+
+    # iterate for all schema
+    for organization in organization_service.get_all(db_session=db_session):
+        schema_engine = engine.execution_options(
+            schema_translate_map={None: f"dispatch_organization_{organization.slug}"}
+        )
+        if not is_scoped:
+            with _session(
+                engine=schema_engine,
+                is_scoped=False,
+            ) as __session:
+                __execute_task_within_session_context(__session)
+        else:
+            with _session(
+                engine=schema_engine,
+                is_scoped=True,
+            ) as __session:
+                __execute_task_within_session_context(__session)
+
+    elapsed_time = time.perf_counter() - start
+    metrics_provider.timer(
+        "function.elapsed.time", value=elapsed_time, tags={"function": fullname(func)}
+    )
+    db_session.close()
+
+
+def scheduled_project_task(func: Callable):
     """Decorator that sets up a background task function with
     a database session and exception tracking.
 
@@ -28,29 +100,31 @@ def scheduled_project_task(func):
 
     @wraps(func)
     def wrapper(*args, **kwargs):
-        db_session = SessionLocal()
-        metrics_provider.counter("function.call.counter", tags={"function": fullname(func)})
-        start = time.perf_counter()
-
-        # iterate for all schema
-        for organization in organization_service.get_all(db_session=db_session):
-            schema_engine = engine.execution_options(
-                schema_translate_map={None: f"dispatch_organization_{organization.slug}"}
-            )
-            schema_session = sessionmaker(bind=schema_engine)()
-            kwargs["db_session"] = schema_session
-            for project in project_service.get_all(db_session=schema_session):
-                kwargs["project"] = project
-                try:
-                    func(*args, **kwargs)
-                except Exception as e:
-                    log.exception(e)
-            schema_session.close()
-        elapsed_time = time.perf_counter() - start
-        metrics_provider.timer(
-            "function.elapsed.time", value=elapsed_time, tags={"function": fullname(func)}
+        _execute_task_in_project_context(
+            func,
+            *args,
+            **kwargs,
+            is_scoped=False,
         )
-        db_session.close()
+
+    return wrapper
+
+
+def scoped_scheduled_project_task(func: Callable):
+    """Decorator that sets up a background task function with
+    a scoped database session and exception tracking.
+
+    Each task is executed in a specific project context.
+    """
+
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        _execute_task_in_project_context(
+            func,
+            *args,
+            **kwargs,
+            is_scoped=True,
+        )
 
     return wrapper