@kevgliss 2023-05-15
I've modified the approach here a little bit. Instead of relying on the scheduler, we have one dedicated thread that processes all new signals in some reasonable order. 

With the other improvements around case creation, this is likely "good enough" for now and we can look into running multiple of these processes concurrently if required (instead of a scheduler doing it for us). 

I also added a`wrk` lua script to aid performance testing.